Feature jonas

+++ lib/AnalogSensor/AnalogSensor.h
@@ -5,20 +5,26 @@
 #include "Sensor.h"
 
 /**
- * Representation of an analog sensor with 1 pin
+ * Representation of an analog sensor with 1 pin,
+ * Implements the Sensor interface.
  */
 class AnalogSensor : public Sensor  {
 private:
     uint8_t pin;
 protected:
-    uint8_t getSensorValue();
+    uint16_t getSensorValue();
 public:
     /**
      * Constructor, creates an instance of an AnalogSensor
-     * 
+     *
      * @param analogPin: The analog pin of the sensor
      */
     AnalogSensor(uint8_t analogPin);
+
+    /**
+     * Abstract method which returns the Sensor value,
+     * Will be implemented in the subclasses
+     */
     virtual float getNormalizedSensorValue() {};
 };
